Update on "add named parameters to mobile module"
[ghstack-poisoned]
