+++ test/cpp/jit/tests.h
@@ -95,6 +95,7 @@ namespace jit {
   _(LiteInterpreterSetState)           \
   _(TorchbindIValueAPI)                \
   _(LiteInterpreterDict)               \
+  _(MobileNamedParameters)             \
   _(FusionAliasing)
 
 #if defined(USE_CUDA)
