+++ torch/csrc/jit/mobile/module.cpp
@@ -100,12 +100,14 @@ void slot_params_recurse(
 void slot_named_params_recurse(
     const c10::intrusive_ptr<c10::ivalue::Object>& obj,
     c10::Dict<std::string, at::Tensor>* params) {
-  for (const auto& slot : obj->slots()) {
+  auto slots = obj->slots();
+  size_t nslots = slots.size();
+  for (size_t i = 0; i < nslots; ++i) {
+    auto slot = slots[i];
     if (slot.isTensor()) {
       auto attributes = obj->type()->getAttributes();
-      for (auto attr : attributes) {
-        params->insert(attr.getName(), slot.toTensor());
-      }
+      auto attr = attributes[i];
+      params->insert(attr.getName(), slot.toTensor());
     } else if (slot.isObject()) {
       slot_named_params_recurse(slot.toObject(), params);
     }
